requirements.txt
flask
psycopg2-binary
python-dotenv
flask-login
flask-bcrypt
